+++ pyxel/run.py
@@ -13,6 +13,8 @@ in MCT, charge diffusion, crosshatches, noises, crosstalk etc.) on a given image
 """
 import argparse
 import logging
+import os
+import shutil
 import sys
 import time
 import typing as t
@@ -266,6 +268,77 @@ def output_directory(configuration: Configuration) -> Path:
     return output_dir
 
 
+def create_model(newmodel: str) -> None:
+    """Create a new module using pyxel/templates/MODELTEMPLATE.py.
+
+    Parameters
+    ----------
+    newmodel: modeltype/modelname
+
+    Returns
+    -------
+    None
+    """
+    try:
+        arguments = newmodel.split("/")
+        location = f"{arguments[0]}"
+        model_name = f"{arguments[1]}"
+    except Exception:
+        sys.exit(
+            f"""
+        Can't create model {arguments}, please use location/newmodelname
+        as an argument for creating a model
+        """
+        )
+
+    # Is not working on UNIX AND Windows if I do not use os.path.abspath
+    path = os.path.abspath(os.getcwd() + "/pyxel/models/" + location + "/")
+    template_string = "_TEMPLATE"
+    template_location = "_LOCATION"
+
+    # Copying the template with the user defined model_name instead
+    src = os.path.abspath(os.getcwd() + "/pyxel/templates/")
+    dest = os.path.abspath(os.getcwd() + "/pyxel/models/" + location + "/")
+    try:
+        try:
+            os.mkdir(dest)
+        except FileExistsError:
+            logging.info(f"{dest} already exists, folder not created")
+        # Replacing all of template in filenames and directories by model_name
+        for dirpath, subdirs, files in os.walk(src):
+            for x in files:
+                pathtofile = os.path.join(dirpath, x)
+                new_pathtofile = os.path.join(
+                    dest, x.replace(template_string, model_name)
+                )
+                shutil.copy(pathtofile, new_pathtofile)
+                # Open file in the created copy
+                with open(new_pathtofile, "r") as file_tochange:
+                    # Replace any mention of template by model_name
+                    new_contents = file_tochange.read().replace(
+                        template_string, model_name
+                    )
+                    new_contents = new_contents.replace(template_location, location)
+                    new_contents = new_contents.replace("%(date)", time.ctime())
+                with open(new_pathtofile, "w+") as file_tochange:
+                    file_tochange.write(new_contents)
+                # Close the file other we can't rename it
+                file_tochange.close()
+
+            for x in subdirs:
+                pathtofile = os.path.join(dirpath, x)
+                os.mkdir(pathtofile.replace(template_string, model_name))
+            logging.info("Module " + model_name + " created.")
+        print("Module " + model_name + " created in " + path + ".")
+    # Directories are the same
+    except shutil.Error as e:
+        logging.critical("Error while duplicating " + template_string + ": %s" % e)
+    # Any error saying that the directory doesn't exist
+    except OSError as e:
+        logging.critical(model_name + " not created. Error: %s" % e)
+    return None
+
+
 def run(input_filename: str, random_seed: t.Optional[int] = None) -> None:
     """TBW.
 
@@ -375,6 +448,15 @@ def main() -> None:
         help="Force flag for saving the examples.",
     )
 
+    parser.add_argument(
+        "-cm",
+        "--createmodel",
+        type=str,
+        help="""Use: -cm arg1/arg2. Create a new module in\
+        pyxel/models/arg1/arg2 using a template\
+        (pyxel/templates/MODELTEMPLATE.py)""",
+    )
+
     # parser.add_argument('-g', '--gui', default=False, type=bool, help='run Graphical User Interface')
     # parser.add_argument('-p', '--port', default=9999, type=int, help='The port to run the web server on')
 
@@ -401,6 +483,8 @@ def main() -> None:
         run(input_filename=opts.config, random_seed=opts.seed)
     elif opts.download_examples:
         download_examples(foldername=opts.download_examples, force=opts.force)
+    elif opts.createmodel:
+        create_model(newmodel=opts.createmodel)
     else:
         print("Define a YAML configuration file!")