+++ tests/unittests/test_calibration.py
@@ -46,7 +46,7 @@ except ImportError:
     ],
 )
 def test_set_algo(yaml):
-    """Test """
+    """Test"""
     cfg = io.load(yaml)
     calibration = cfg.calibration
     obj = calibration.algorithm.get_algorithm()
@@ -86,7 +86,7 @@ def test_set_algo(yaml):
     ],
 )
 def test_read_data(input_data):
-    """Test """
+    """Test"""
     output = read_data(input_data)
     assert isinstance(output, list)
 
@@ -137,7 +137,7 @@ def test_read_single_data(input_data):
     ],
 )
 def test_list_to_slice(input_data):
-    """Test """
+    """Test"""
     output = list_to_slice(input_data)
     if isinstance(output, slice):
         pass
@@ -171,7 +171,7 @@ def test_list_to_slice(input_data):
     ],
 )
 def test_check_ranges(targ_range, out_range, row, col):
-    """Test """
+    """Test"""
     with pytest.raises(ValueError):
         check_ranges(targ_range, out_range, row, col)
 
@@ -180,7 +180,7 @@ def test_check_ranges(targ_range, out_range, row, col):
 @pytest.mark.skipif(not WITH_PYGMO, reason="Package 'pygmo' is not installed.")
 @pytest.mark.parametrize("yaml", ["tests/data/calibrate_models.yaml"])
 def test_run_calibration(yaml):
-    """Test """
+    """Test"""
     cfg = io.load(yaml)
     assert isinstance(cfg, Configuration)