+++ tests/unittests/test_fitting.py
@@ -51,7 +51,7 @@ def configure(mf: ModelFitting, sim: Configuration) -> None:
 @pytest.mark.skipif(not WITH_PYGMO, reason="Package 'pygmo' is not installed.")
 @pytest.mark.parametrize("yaml_file", ["tests/data/calibrate.yaml"])
 def test_configure_params(yaml_file):
-    """Test """
+    """Test"""
     cfg = io.load(yaml_file)
     detector = cfg.ccd_detector
     pipeline = cfg.pipeline
@@ -73,7 +73,7 @@ def test_configure_params(yaml_file):
 @pytest.mark.skipif(not WITH_PYGMO, reason="Package 'pygmo' is not installed.")
 @pytest.mark.parametrize("yaml", ["tests/data/calibrate_fits.yaml"])
 def test_configure_fits_target(yaml):
-    """Test """
+    """Test"""
     cfg = io.load(yaml)
     detector = cfg.ccd_detector
     pipeline = cfg.pipeline
@@ -101,7 +101,7 @@ def test_configure_fits_target(yaml):
     "yaml", ["tests/data/calibrate.yaml", "tests/data/calibrate_fits.yaml"]
 )
 def test_boundaries(yaml):
-    """Test """
+    """Test"""
     cfg = io.load(yaml)
     detector = cfg.ccd_detector
     pipeline = cfg.pipeline